prime table tests use long arrays

+++ test/primetables/PrimeTablesTest.java
@@ -25,8 +25,8 @@ public class PrimeTablesTest {
     public void testGeneratePrimeTable() {
         int[] primes = {2, 3, 5};
         int rowNo = 1;
-        int[] expResult = {3, 6, 9, 15};
-        int[] result = PrimeTables.generatePrimeTableRow(primes, rowNo);
+        long[] expResult = {3, 6, 9, 15};
+        long[] result = PrimeTables.generatePrimeTableRow(primes, rowNo);
         
         System.out.println("generatePrimeTable");
